+++ src/main/java/com/gerarecibos/recibos/DTO/ProdutoDto.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class ProdutoDto {
     private String produtoNome;
-    private Double produtoValorTotal;
+    private Integer produtoValorTotal;
     private String produtoDescricao; // Novo campo para descrição
 }
